GVT-2387 Move LocalizationKey into localization package
