+++ infra/src/main/kotlin/fi/fta/geoviite/infra/error/ApiErrorResponse.kt
@@ -1,7 +1,7 @@
 package fi.fta.geoviite.infra.error
 
 import fi.fta.geoviite.infra.localization.LocalizationParams
-import fi.fta.geoviite.infra.util.LocalizationKey
+import fi.fta.geoviite.infra.localization.LocalizationKey
 import java.time.Instant
 
 data class ApiErrorResponse(
