+++ infra/src/main/kotlin/fi/fta/geoviite/infra/error/ErrorHandling.kt
@@ -10,7 +10,7 @@ import com.fasterxml.jackson.databind.exc.MismatchedInputException
 import com.fasterxml.jackson.databind.exc.ValueInstantiationException
 import fi.fta.geoviite.infra.localization.LocalizationParams
 import fi.fta.geoviite.infra.localization.localizationParams
-import fi.fta.geoviite.infra.util.LocalizationKey
+import fi.fta.geoviite.infra.localization.LocalizationKey
 import jakarta.xml.bind.UnmarshalException
 import org.geotools.api.referencing.operation.TransformException
 import org.postgresql.util.PSQLException
