+++ infra/src/main/kotlin/fi/fta/geoviite/infra/inframodel/InfraModelParsing.kt
@@ -10,7 +10,7 @@ import fi.fta.geoviite.infra.geometry.GeometryValidationIssue
 import fi.fta.geoviite.infra.switchLibrary.SwitchStructure
 import fi.fta.geoviite.infra.switchLibrary.SwitchType
 import fi.fta.geoviite.infra.util.FileName
-import fi.fta.geoviite.infra.util.LocalizationKey
+import fi.fta.geoviite.infra.localization.LocalizationKey
 import fi.fta.geoviite.infra.util.xmlBytesToString
 import jakarta.xml.bind.JAXBContext
 import jakarta.xml.bind.Marshaller
